STY: Make classess camelcase, remove name kwarg, move into lasers folder.
